Add checks

+++ Ontwikkelopdracht.Persistence/RepositoryArmour.cs
@@ -35,16 +35,35 @@ namespace Ontwikkelopdracht.Persistence
 
         public void Delete(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(id), id, "Valid values are x >= 0.");
+            }
+
             Repository.Delete(id);
         }
 
         public void Delete(List<T> entities)
         {
+            if (entities == null)
+            {
+                throw new ArgumentNullException(nameof(entities));
+            }
+            if (entities.Any(entity => entity == null))
+            {
+                throw new ArgumentNullException(nameof(entities), "The list contains a null reference.");
+            }
+
             Repository.Delete(entities);
         }
 
         public void Delete(T entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+
             Repository.Delete(entity);
         }
 
@@ -55,6 +74,11 @@ namespace Ontwikkelopdracht.Persistence
 
         public bool Exists(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(id), id, "Valid values are x >= 0.");
+            }
+
             return Repository.Exists(id);
         }
 
@@ -65,31 +89,65 @@ namespace Ontwikkelopdracht.Persistence
 
         public List<T> FindAll(List<int> ids)
         {
+            if (ids == null)
+            {
+                throw new ArgumentNullException(nameof(ids));
+            }
+
             return Repository.FindAll(ids);
         }
 
         public List<T> FindAllWhere(Func<T, bool> predicate)
         {
+            if (predicate == null)
+            {
+                throw new ArgumentNullException(nameof(predicate));
+            }
+
             return Repository.FindAllWhere(predicate);
         }
 
         public List<T> FindAllWhere(Func<T, int, bool> predicate)
         {
+            if (predicate == null)
+            {
+                throw new ArgumentNullException(nameof(predicate));
+            }
+
             return Repository.FindAllWhere(predicate);
         }
 
         public T FindOne(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(id), id, "Valid values are x >= 0.");
+            }
+
             return Repository.FindOne(id);
         }
 
         public T Save(T entity)
         {
+            if (entity == null)
+            {
+                throw new ArgumentNullException(nameof(entity));
+            }
+
             return Repository.Save(entity);
         }
 
         public List<T> Save(List<T> entities)
         {
+            if (entities == null)
+            {
+                throw new ArgumentNullException(nameof(entities));
+            }
+            if (entities.Any(entity => entity == null))
+            {
+                throw new ArgumentNullException(nameof(entities), "The list contains a null reference.");
+            }
+
             return Repository.Save(entities);
         }
     }
